refactor: migrate recommendation_refiner to drift-output
(cherry picked from commit 5d4490435d035795f74b832b2aed0cdc94af7f35)
